Refactor interbit tests for parity with src folder structure
